+++ lib/camera/camera.dart
@@ -273,7 +273,7 @@ class Camera with HasGameRef<BonfireGame> {
   }
 
   Offset _getCenterTarget() {
-    if (config.target is ObjectCollision) {
+    if (config.target?.isObjectCollision() == true) {
       return (config.target as ObjectCollision).rectCollision.center;
     }
     return config.target?.position.rect.center ?? Offset.zero;