+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.3
+- Add type `above` in tiled to render above components
+- update `tiledjsonreader`
+- improvements in `Camera`
+- add param `objectsBuilder` in `TiledWorldMap`
+- others improvements
+
 # 1.0.2
 - Downgrade flame version to 1.0.0-rc9